Move `enum Token` to `refs/mod.rs`
We leave the tests associated with `enum Token` in a separate source
file which we rename to `refs/token_tests.rs`. Additionally, we move the
tests for the `parse_ref()` function to a separate source file.
